+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v2.1.3] - 2017-01-18
 - Date added on every page of the report, major mutations are in bold.
 - Bugfix: minvar now works evene when neither HCV nor HIV reads are present.
 - Bugfix: `find_subtype` failed when reads from both HIV and HCV were present.